Project for CIS 3050
A database project involving MS SQL. This MS SQL project introduces various aspects of the SQL SELECT statement and the
methods of retrieving data from the database tables. The project also introduces the
fundamentals of updating and deleting records. It utilizes a set of tables that are
represented by the ERD and are created and populated by the script file.